feat: nested routes for package tabs
